+++ musefs-core/tests/interop_emit.rs
@@ -148,8 +148,8 @@ fn emit(
     dst: &Path,
     bytes: &[u8],
     format: Format,
-    audio_offset: i64,
-    audio_length: i64,
+    audio_offset: u64,
+    audio_length: u64,
     arts: &[(&[u8], &str)],
 ) -> (u64, u64) {
     std::fs::write(src, bytes).unwrap();
@@ -160,7 +160,7 @@ fn emit(
             format,
             audio_offset,
             audio_length,
-            backing_size: std::fs::metadata(src).unwrap().len() as i64,
+            backing_size: std::fs::metadata(src).unwrap().len(),
             backing_mtime: real_mtime(src),
         })
         .unwrap();
@@ -210,8 +210,8 @@ fn emit_binary(
     dst: &Path,
     bytes: &[u8],
     format: Format,
-    audio_offset: i64,
-    audio_length: i64,
+    audio_offset: u64,
+    audio_length: u64,
     text: &[Tag],
     binary: &[BinaryTag],
 ) {
@@ -223,7 +223,7 @@ fn emit_binary(
             format,
             audio_offset,
             audio_length,
-            backing_size: std::fs::metadata(src).unwrap().len() as i64,
+            backing_size: std::fs::metadata(src).unwrap().len(),
             backing_mtime: real_mtime(src),
         })
         .unwrap();
@@ -251,8 +251,8 @@ fn emit_interop_fixtures() {
             &dir.join("out.flac"),
             &bytes,
             Format::Flac,
-            scan.audio_offset as i64,
-            scan.audio_length as i64,
+            scan.audio_offset,
+            scan.audio_length,
             &[],
         );
         manifest.push(ManifestRow {
@@ -278,8 +278,8 @@ fn emit_interop_fixtures() {
             &dir.join("out.mp3"),
             &bytes,
             Format::Mp3,
-            b.audio_offset as i64,
-            b.audio_length as i64,
+            b.audio_offset,
+            b.audio_length,
             &[],
         );
         manifest.push(ManifestRow {
@@ -306,8 +306,8 @@ fn emit_interop_fixtures() {
             &dir.join("out.m4a"),
             &bytes,
             Format::M4a,
-            scan.mdat_payload_offset as i64,
-            scan.mdat_payload_len as i64,
+            scan.mdat_payload_offset,
+            scan.mdat_payload_len,
             &[(COVR_JPEG, "image/jpeg"), (COVR_PNG, "image/png")],
         );
         manifest.push(ManifestRow {
@@ -333,8 +333,8 @@ fn emit_interop_fixtures() {
             &dir.join("out.ogg"),
             &bytes,
             Format::Opus,
-            scan.audio_offset as i64,
-            scan.audio_length as i64,
+            scan.audio_offset,
+            scan.audio_length,
             &[],
         );
         manifest.push(ManifestRow {
@@ -360,8 +360,8 @@ fn emit_interop_fixtures() {
             &dir.join("out.wav"),
             &bytes,
             Format::Wav,
-            b.audio_offset as i64,
-            b.audio_length as i64,
+            b.audio_offset,
+            b.audio_length,
             &[],
         );
         manifest.push(ManifestRow {
@@ -406,8 +406,8 @@ fn emit_interop_fixtures() {
             &dir.join("out_bin.mp3"),
             &bytes,
             Format::Mp3,
-            b.audio_offset as i64,
-            b.audio_length as i64,
+            b.audio_offset,
+            b.audio_length,
             &[
                 Tag::new("title", "Bin Title", 0),
                 Tag::new("artist", "Bin Artist", 0),
@@ -439,8 +439,8 @@ fn emit_interop_fixtures() {
             &dir.join("out_bin.m4a"),
             &bytes,
             Format::M4a,
-            scan.mdat_payload_offset as i64,
-            scan.mdat_payload_len as i64,
+            scan.mdat_payload_offset,
+            scan.mdat_payload_len,
             &[
                 Tag::new("title", "Bin Title", 0),
                 Tag::new("artist", "Bin Artist", 0),
